chore: move getTimeLeft to utils
